Before:

```
Get performance: 4194304 items in 13.99276754213497 seconds
Get performance multi-thread: 4194304 items in 13.610933708958328 seconds
```

After:

```
Get performance: 4194304 items in 14.377041833940893 seconds
Get performance multi-thread: 4194304 items in 5.503023374825716 seconds
```