@Congyuwang 2023-12-13
So, the improvement is mainly on batch_get right? Since batch_get usually needs more time. 👍