@GodTamIt 2023-12-13
Right, batch gets can run for an extended amount of time without having to re-acquire the GIL to access Python objects.

Individual `get()`s do not because they need the GIL to access the Python object values for the key and value every time. So if you do a get for a bunch of items, you're going to be contending on the lock for each iteration.