@GodTamIt 2023-12-13
For posterity, I'll comment here a few more details:

- The reason we don't see a 4x performance increase here is because the benchmark iterates through the results to verify them (which requires the GIL).
- The numbers were run on a M2 MacBook Pro 16.